split downloader to its own file
